requirements.txt
django
mysqlclient
Pillow
djangorestframework
requests
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
django-jazzmin
django-js-asset
Jinja2
redis
template
bcrypt

PyMuPDF
reportlab
pdf2docx
django-environ
requests 
django-cors-headers
celery
django-celery-results
concurrent-log-handler
beautifulsoup4
drf-spectacular

# pip install -r requirements.txt
# python manage.py migrate
# python manage.py runserver


# root@063318baf6e9:/# ^C


# apt install -y libjpeg-dev zlib1g-dev libfreetype6-dev liblcms2-dev libopenjpeg-dev libtiff-dev libwebp-dev